+++ warcrecord/warcfieldsblock.go
@@ -51,7 +51,7 @@ func NewWarcFieldsBlock(block Block, options *warcoptions.WarcOptions) (WarcFiel
 	return &warcFieldsBlock{Block: block, warcFields: wf.(warcfields.WarcFields)}, nil
 }
 
-func (b *warcFieldsBlock) Write(w io.Writer) (bytesWritten int, err error) {
+func (b *warcFieldsBlock) Write(w io.Writer) (bytesWritten int64, err error) {
 	bytesWritten, err = b.warcFields.Write(w)
 	w.Write([]byte(CRLF))
 	bytesWritten += 2
